# ARROW-2402: [C++] Avoid spurious copies with FixedSizeBinaryBuilder