The lack of a FixedSizeBinaryBuilder::Append() overload for "const char*" implied a temporary conversion to std::string.