@pitrou 2018-04-05
Before:
```
Benchmark                                              Time           CPU Iterations
------------------------------------------------------------------------------------
BM_BuildFixedSizeBinaryArray/repeats:3             88124 us      88090 us          8    113.52MB/s
BM_BuildFixedSizeBinaryArray/repeats:3             87680 us      87644 us          8   114.098MB/s
BM_BuildFixedSizeBinaryArray/repeats:3             88302 us      88262 us          8   113.299MB/s
BM_BuildFixedSizeBinaryArray/repeats:3_mean        88036 us      87999 us          8   113.639MB/s
BM_BuildFixedSizeBinaryArray/repeats:3_stddev        262 us        261 us          0   345.013kB/s
```

After:
```
Benchmark                                              Time           CPU Iterations
------------------------------------------------------------------------------------
BM_BuildFixedSizeBinaryArray/repeats:3             12534 us      12506 us         56   799.587MB/s
BM_BuildFixedSizeBinaryArray/repeats:3             12553 us      12524 us         56   798.455MB/s
BM_BuildFixedSizeBinaryArray/repeats:3             12537 us      12518 us         56    798.85MB/s
BM_BuildFixedSizeBinaryArray/repeats:3_mean        12541 us      12516 us         56   798.964MB/s
BM_BuildFixedSizeBinaryArray/repeats:3_stddev          8 us          7 us          0   480.683kB/s
```