@pitrou 2018-04-05
Also noticeable in Python (the "binary10" row):
* before:
```
[100.00%] ··· Running convert_builtins.ConvertPyListToArray.time_convert                                                                                 ok
[100.00%] ···· 
               ==================== =============
                       type                      
               -------------------- -------------
                      int32          5.67±0.08ms 
                      uint32         5.32±0.06ms 
                      int64           5.71±0.1ms 
                      uint64          5.44±0.1ms 
                     float32         4.77±0.04ms 
                     float64         5.04±0.04ms 
                       bool          4.13±0.01ms 
                     decimal          180±0.3ms  
                      binary          8.77±0.1ms 
                     binary10         12.2±0.2ms 
                      ascii           10.5±0.2ms 
                     unicode          18.4±0.7ms 
                    int64 list        52.2±0.6ms 
                      struct           32.5±1ms  
                struct from tuples     32.0±2ms  
               ==================== =============
```
* after:
```
[100.00%] ··· Running convert_builtins.ConvertPyListToArray.time_convert                                                                                 ok
[100.00%] ···· 
               ==================== =============
                       type                      
               -------------------- -------------
                      int32           5.65±0.1ms 
                      uint32          5.26±0.1ms 
                      int64           5.90±0.1ms 
                      uint64          5.76±0.1ms 
                     float32         5.23±0.03ms 
                     float64         5.42±0.04ms 
                       bool          4.06±0.02ms 
                     decimal          183±0.4ms  
                      binary          8.62±0.1ms 
                     binary10         5.78±0.1ms 
                      ascii           10.4±0.2ms 
                     unicode          18.2±0.8ms 
                    int64 list        50.6±0.7ms 
                      struct           32.5±1ms  
                struct from tuples     31.7±1ms  
               ==================== =============
```